refactor: more files migrated
