+++ gradle/libs.versions.toml
@@ -1,4 +1,4 @@
 [versions]
-revanced-patcher = "19.3.1"
-revanced-patches = "4.8.0-dev.8"
-smali = "3.0.5"
+revanced-patcher = "20.0.0-dev.1"
+# revanced-patches = "4.8.0-dev.8"
+smali = "3.0.7"
