+++ patches/src/main/kotlin/dropped/patches/fxfileexplorer/unlockplus/unlockPlusPatch.kt
@@ -11,6 +11,7 @@ val unlockPlusPatch = bytecodePatch(
     compatibleWith("nextapp.fx")
 
     val unlockPlusPatch by isPlusUnlockedFingerprint()
+
     execute { 
       unlockPlusPatch.mutableMethod.addInstructions(
         0,
