+++ patches/src/main/kotlin/dropped/patches/nova/prime/UnlockPrimeFingerprint.kt
@@ -3,7 +3,7 @@ package dropped.patches.nova.prime
 import app.revanced.patcher.fingerprint
 import com.android.tools.smali.dexlib2.Opcode
 
-internal val updatePrimeStatusFingerprint = fingerprint {
-    opcodes(Opcode.CONST_4), // Modify this to change the prime status.
+internal val unlockPrimeFingerprint = fingerprint {
+    opcodes(Opcode.CONST_4) // Modify this to change the prime status.
     strings("widget_reset_ids")
   }
